make utils into a separate package
