+++ packages/utils/README.md
@@ -1,0 +1,1 @@
+# @solid-primitives/utils
