+++ packages/utils/tsconfig-cjs.json
@@ -1,0 +1,9 @@
+{
+	"extends": "./tsconfig.json",
+	"compilerOptions": {
+		"module": "CommonJS",
+		"target": "ES6",
+		"outDir": "./dist/cjs",
+		"declaration": false
+	}
+}
